refactor(shelves): extract ShelfPlacementCoordinator from GpuStorePropsRenderer
